requirements.txt
datasketch>=1.5
xxhash>=3.4
tqdm>=4.64
numpy>=1.23
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3
ruff>=0.1
black>=23.7
pytest>=7.4
pyyaml>=6.0
plotly>=5.18
sentence-transformers>=2.2
hypothesis>=6.88
spacy>=3.6
tf-keras>=2.15
joblib>=1.3
psutil>=5.9
langdetect>=1.0.9
beautifulsoup4>=4.12
presidio-analyzer>=2.2
presidio-anonymizer>=2.2
transformers>=4.30
pyarrow>=14.0
fastparquet>=2023.10
chardet>=5.2
lxml>=4.9
detoxify>=0.5.2 